refactor: board 패키지 일부 클래스 piece 내부로 이동
- (결정기록) 패키지 순환 참조를 해결하기 위해서, 이동시켰다.
