+++ src/main/java/chess/application/request/MovePieceRequest.java
@@ -1,6 +1,6 @@
 package chess.application.request;
 
-import chess.domain.board.Coordinate;
+import chess.domain.piece.Coordinate;
 
 public record MovePieceRequest(Coordinate source, Coordinate target) {
 }
