+++ src/main/java/chess/domain/board/Board.java
@@ -1,7 +1,10 @@
 package chess.domain.board;
 
 import java.util.Map;
+import chess.domain.piece.Coordinate;
 import chess.domain.piece.Piece;
+import chess.domain.piece.Pieces;
+import chess.domain.piece.PiecesFactory;
 import chess.domain.piece.Team;
 
 public class Board {
