+++ src/main/java/chess/domain/piece/King.java
@@ -1,8 +1,6 @@
 package chess.domain.piece;
 
 import java.util.List;
-import chess.domain.board.Coordinate;
-import chess.domain.board.Pieces;
 
 public class King extends AbstractNonSlidingPiece {
