+++ src/main/java/chess/domain/piece/Piece.java
@@ -1,8 +1,5 @@
 package chess.domain.piece;
 
-import chess.domain.board.Coordinate;
-import chess.domain.board.Pieces;
-
 public interface Piece {
 
     void validateMovable(Coordinate source, Coordinate target, Pieces pieces);
