+++ src/main/java/chess/domain/piece/PiecesFactory.java
@@ -1,15 +1,7 @@
-package chess.domain.board;
+package chess.domain.piece;
 
 import java.util.HashMap;
 import java.util.List;
-import chess.domain.piece.Bishop;
-import chess.domain.piece.King;
-import chess.domain.piece.Knight;
-import chess.domain.piece.Pawn;
-import chess.domain.piece.Piece;
-import chess.domain.piece.Queen;
-import chess.domain.piece.Rook;
-import chess.domain.piece.Team;
 
 public class PiecesFactory {
