+++ src/test/java/chess/domain/piece/FileTest.java
@@ -1,4 +1,4 @@
-package chess.domain.board;
+package chess.domain.piece;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -8,20 +8,20 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-class RankTest {
+class FileTest {
 
     @DisplayName("생성 테스트")
     @Test
     void create() {
-        assertThatCode(() -> new Rank(1))
+        assertThatCode(() -> new File('a'))
                 .doesNotThrowAnyException();
     }
 
-    @DisplayName("행은 1부터 8사이의 값만 허용한다.")
+    @DisplayName("열은 소문자 a부터 h까지의 값만 허용한다.")
     @ParameterizedTest
-    @ValueSource(ints = {0, 9})
-    void checkValue(int source) {
-        assertThatThrownBy(() -> new Rank(source))
+    @ValueSource(chars = {'A', 'i'})
+    void checkValue(char source) {
+        assertThatThrownBy(() -> new File(source))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 }
